Keep cost distribution lookup ID as int

+++ src/ERPAccounting.Domain/Lookups/LookupModels.cs
@@ -92,7 +92,7 @@ public record CostTypeLookup(
 public record CostDistributionMethodLookup
 {
     [Column("IDNacinDeljenjaTroskova")]
-    public short IdNacinDeljenjaTroskova { get; set; }
+    public int IdNacinDeljenjaTroskova { get; set; }
 
     [Column("Naziv")]
     public string Naziv { get; set; } = string.Empty;
